make test more illustrative of how to use meta plus dtype param in gufunc

+++ dask/array/tests/test_gufunc.py
@@ -620,10 +620,10 @@ def test_preserve_meta_type():
 
 def test_apply_gufunc_with_meta():
     def stats(x):
-        return np.mean(x, axis=-1), np.std(x, axis=-1)
+        return np.mean(x, axis=-1), np.std(x, axis=-1, dtype=np.float32)
 
     a = da.random.normal(size=(10, 20, 30), chunks=(5, 5, 30))
-    meta = (np.ones(0, dtype=np.float64), np.ones(0, dtype=np.float64))
+    meta = (np.ones(0, dtype=np.float64), np.ones(0, dtype=np.float32))
     result = apply_gufunc(stats, "(i)->(),()", a, meta=meta)
     expected = stats(a.compute())
     assert_eq(expected[0], result[0])
